GA  1.0
Right we are just  generating bodies and rewards based on the number of
parts.

Starts with a seed population all containing  a single part of dna.  Two
offspring are introduced from the general population

When the new scores  are settled camera tracks the highest ranking
member.
and from the  elite set.